+++ lib/fbutton.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:fcontrol/fcontrol.dart';
 
 /// [FButton] 图片相对与文字的位置
 ///
@@ -8,14 +9,17 @@ enum ImageAlignment {
   ///
   /// Left of text
   left,
+
   /// 文字顶部
   ///
   /// Top of text
   top,
+
   /// 文字右边
   ///
   /// Right side of text
   right,
+
   /// 文字底部
   ///
   /// Bottom of text
@@ -46,242 +50,250 @@ class FButton extends StatefulWidget {
   /// 当按下按钮时，会触发该函数。如果未设置该函数，按钮将进入不可用状态。
   ///
   /// When the button is pressed, the function is triggered.If this function is not set, the button will enter an unavailable state.
-  VoidCallback onPressed;
+  final VoidCallback onPressed;
 
   /// 按钮上的文字。
   ///
   /// The text on the button.
-  String text;
+  final String text;
+
+  /// 按钮文本样式
+  ///
+  /// Button text style
+  final TextStyle style;
+
+  final Alignment alignment;
 
   /// 按钮文字的颜色。
   ///
   /// The color of the button text.
-  Color textColor;
+  @deprecated
+  final Color textColor;
 
   /// 按钮文字的大小
   ///
   /// Button text size
-  double fontSize;
+  @deprecated
+  final double fontSize;
 
   /// 按钮文字的风格
   ///
   /// Button text style
-  FontStyle fontStyle;
+  @deprecated
+  final FontStyle fontStyle;
 
   /// 一行文字高度 = textSize * fontHeight
   ///
   /// height of one line of text = textSize * fontHeight
-  double fontHeight;
+  @deprecated
+  final double fontHeight;
 
   /// 文字粗细。默认 [FontWeight.normal]
   ///
   /// Text weight. Default [FontWeight.normal]
-  FontWeight textWeight;
+  @deprecated
+  final FontWeight fontWeight;
 
   /// 按钮文字的不可用颜色
   ///
   /// Unavailable colors for button text
-  Color disabledTextColor;
+  final Color disabledTextColor;
 
   /// 按钮的颜色
   ///
   /// Button color
-  Color color;
+  final Color color;
 
   /// 按钮不可用颜色
   ///
   /// Button is not available in color
-  Color disabledColor;
+  final Color disabledColor;
 
   /// 当 [effect] 为 true 时，触摸按钮时涟漪的颜色。
   ///
   /// When [effect] is true, the color of the ripple when the button is touched.
-  Color splashColor;
+  final Color splashColor;
 
   /// 当 [effect] 为 true 时，鼠标进入按钮范围时，按钮的颜色。
   ///
   /// When [effect] is true, the color of the button when the mouse enters the button range.
-  Color hoverColor;
+  final Color hoverColor;
 
   /// 当 [effect] 为 true 时，按压按钮时的按钮颜色。
   ///
   /// When [effect] is true, the button color when the button is pressed.
-  Color highlightColor;
+  final Color highlightColor;
 
   /// 内间距。
   ///
   /// Internal spacing
-  EdgeInsetsGeometry padding;
+  final EdgeInsetsGeometry padding;
 
   /// 焦点。
   ///
   /// focus
-  FocusNode focusNode;
+  final FocusNode focusNode;
 
   /// 是否能够自动获取焦点
   ///
   /// Whether it can automatically obtain focus
-  bool autofocus;
+  final bool autofocus;
 
   /// 宽度。
   ///
   /// width
-  double width;
+  final double width;
 
   /// 高度。
   ///
   /// height.
-  double height;
+  final double height;
 
   /// 设置组件圆角。详见 [FButtonCorner]
   ///
   /// Sets the component fillet. See [FButtonCorner] for details
-  FButtonCorner corner;
+  final FButtonCorner corner;
 
   /// 设置圆角风格，默认 [FButtonCorner.round]
   ///
   /// Set rounded corner style, default [FButtonCorner.round]
-  FButtonCornerStyle cornerStyle;
+  final FButtonCornerStyle cornerStyle;
 
   /// 设置边框颜色。
   ///
   /// Set the border color.
-  Color strokeColor;
+  final Color strokeColor;
 
   /// 设置边框宽
   ///
   /// Set border width
-  double strokeWidth;
+  final double strokeWidth;
 
   /// 设置组件阴影颜色
   ///
   /// Set component shadow color
-  Color shadowColor;
+  final Color shadowColor;
 
   /// 设置组件阴影偏移
   ///
   /// Set component shadow offset
-  Offset shadowOffset;
+  final Offset shadowOffset;
 
   /// 设置组件高斯与阴影形状卷积的标准偏差。
   ///
   /// Sets the standard deviation of the component's Gaussian convolution with the shadow shape.
-  double shadowBlur;
+  final double shadowBlur;
 
   /// 是否开启特效。开启特效后，[splashColor]、[highlightColor]、[hoverColor] 等属性才有效。
   ///
   /// Whether to enable special effects. After the special effects are turned on, the attributes such as [splashColor], [highlightColor], [hoverColor] are effective.
-  bool effect;
+  final bool effect;
 
   /// 是否开启基于阴影的点击特效。
   ///
   /// Whether to enable shadow-based click effects.
-  bool clickEffect;
+  final bool clickEffect;
 
   /// 设置组件渐变色背景。会覆盖 [color] 配置
   /// 你可选择 [LinearGradient]，[RadialGradient]，[SweepGradient] 等..
   ///
   /// Sets the gradient background of the component. [BackgroundColor]
   /// You can choose [LinearGradient], [RadialGradient], [SweepGradient], etc ..
-  Gradient gradient;
+  final Gradient gradient;
 
   /// 设置图标
   ///
   /// Settings icon
-  Widget image;
+  final Widget image;
 
   /// 设置图标与文本的间距
   ///
   /// Set the distance between the icon and the text
-  double imageMargin;
+  final double imageMargin;
 
   /// 设置图标与文本的相对位置。详见 [ImageAlignment]
   ///
   /// Set the relative position of the icon and the text. See [ImageAlignment] for details
-  ImageAlignment imageAlignment;
+  final ImageAlignment imageAlignment;
 
   /// 是否启动 Loading 状态。Loading 状态会覆盖 [image] 配置
   ///
   /// Whether to start the loading state. Loading status will override [image] configuration
-  bool loading;
+  final bool loading;
 
   /// Loading 的颜色
   ///
   /// Loading colors
-  Color loadingColor;
+  final Color loadingColor;
 
   /// Loading 的宽度
   ///
   /// Loading width
-  double loadingStrokeWidth;
+  final double loadingStrokeWidth;
 
   /// Loading 的大小
   ///
   /// Loading size
-  double loadingSize;
+  final double loadingSize;
 
   /// 是否启用点击进入 Loading 状态的模式
   ///
   /// Whether to enable click to enter the loading mode
-  bool clickLoading;
+  final bool clickLoading;
 
   /// Loading 状态下是否隐藏文本
   ///
   /// Whether to hide text in the loading state
-  bool hideTextOnLoading;
+  final bool hideTextOnLoading;
 
   /// Loading 状态下展示的文本
   ///
   /// Text displayed under Loading
-  String loadingText;
+  final String loadingText;
 
-  double cacheShadowBlur;
-
-  Color cacheShadowColor;
-
-  bool get enabled => onPressed != null && !loading;
-
-  FButton(
-      {Key key,
-      @required this.onPressed,
-      this.text,
-      this.textColor,
-      this.disabledTextColor,
-      this.fontSize,
-      this.fontStyle,
-      this.color,
-      this.disabledColor,
-      this.hoverColor,
-      this.highlightColor,
-      this.splashColor,
-      this.effect = false,
-      this.padding,
-      this.focusNode,
-      this.autofocus = false,
-      this.width,
-      this.height,
-      this.corner,
-      this.cornerStyle = FButtonCornerStyle.round,
-      this.strokeColor,
-      this.strokeWidth,
-      this.shadowColor,
-      this.shadowOffset,
-      this.shadowBlur = 1.0,
-      this.gradient,
-      this.image,
-      this.imageMargin = 6.0,
-      this.imageAlignment = ImageAlignment.left,
-      this.loading = false,
-      this.loadingColor,
-      this.loadingStrokeWidth = 4.0,
-      this.clickLoading = false,
-      this.hideTextOnLoading = false,
-      this.loadingText,
-      this.loadingSize = 12,
-      this.clickEffect = false})
-      : cacheShadowBlur = shadowBlur,
-        cacheShadowColor = shadowColor,
-        assert(autofocus != null),
+  FButton({
+    Key key,
+    this.onPressed,
+    this.text,
+    this.textColor,
+    this.disabledTextColor,
+    this.fontSize,
+    this.fontStyle,
+    this.color,
+    this.disabledColor,
+    this.hoverColor,
+    this.highlightColor,
+    this.splashColor,
+    this.effect = false,
+    this.padding,
+    this.focusNode,
+    this.autofocus = false,
+    this.width,
+    this.height,
+    this.corner,
+    this.cornerStyle = FButtonCornerStyle.round,
+    this.strokeColor,
+    this.strokeWidth,
+    this.shadowColor,
+    this.shadowOffset,
+    this.shadowBlur = 1.0,
+    this.gradient,
+    this.image,
+    this.imageMargin = 6.0,
+    this.imageAlignment = ImageAlignment.left,
+    this.loading = false,
+    this.loadingColor,
+    this.loadingStrokeWidth = 4.0,
+    this.clickLoading = false,
+    this.hideTextOnLoading = false,
+    this.loadingText,
+    this.loadingSize = 12,
+    this.clickEffect = false,
+    this.style,
+    this.alignment,
+    this.fontHeight,
+    this.fontWeight,
+  })  : assert(autofocus != null),
         super(key: key);
 
   @override
@@ -293,35 +305,57 @@ class FButton extends StatefulWidget {
 class _FButton extends State<FButton> {
   final Set<MaterialState> _states = <MaterialState>{};
 
+  bool effect;
+
+  double shadowBlur;
+
+  double cacheShadowBlur;
+
+  Color shadowColor;
+
+  Color cacheShadowColor;
+
   bool get _hovered => _states.contains(MaterialState.hovered);
 
   bool get _pressed => _states.contains(MaterialState.pressed);
 
   bool get _disabled => _states.contains(MaterialState.disabled);
 
-  void _updateState(MaterialState state, bool value) {
+  bool get enabled => widget.onPressed != null && !loading;
+
+  bool loading = false;
+
+  void updateState(MaterialState state, bool value) {
     value ? _states.add(state) : _states.remove(state);
   }
 
   @override
   void initState() {
     super.initState();
-    _updateState(MaterialState.disabled, !widget.enabled);
+    updateParam();
+    updateState(MaterialState.disabled, !enabled);
   }
 
   @override
   void didUpdateWidget(FButton oldWidget) {
     super.didUpdateWidget(oldWidget);
-    _updateState(MaterialState.disabled, !widget.enabled);
+    updateParam();
+    updateState(MaterialState.disabled, !enabled);
     if (_disabled && _pressed) {
       _handleHighlightChanged(false);
     }
   }
 
+  updateParam() {
+    shadowBlur = widget.shadowBlur;
+    shadowColor = widget.shadowColor;
+    loading = widget.loading;
+  }
+
   void _handleHighlightChanged(bool value) {
     if (_pressed != value) {
       setState(() {
-        _updateState(MaterialState.pressed, value);
+        updateState(MaterialState.pressed, value);
       });
     }
   }
@@ -329,28 +363,29 @@ class _FButton extends State<FButton> {
   void _handleHoveredChanged(bool value) {
     if (_hovered != value) {
       setState(() {
-        _updateState(MaterialState.hovered, value);
+        updateState(MaterialState.hovered, value);
       });
     }
   }
 
   @override
   Widget build(BuildContext context) {
+    cacheShadowBlur = widget.shadowBlur;
+    cacheShadowColor = widget.shadowColor;
     var themeData = Theme.of(context);
     final ButtonThemeData buttonTheme = themeData.buttonTheme;
     final TextStyle textStyle = themeData.textTheme.button;
-    bool gradientEnable = widget.enabled && widget.gradient != null;
+    bool gradientEnable = enabled && widget.gradient != null;
     double disableOpacity = 0.68;
 
     /// determine whether to enable special effects
-    widget.effect = widget.enabled ? widget.effect : false;
-    widget.color = widget.enabled
-        ? widget.color ?? buttonTheme.colorScheme.background
-        : widget.disabledColor ??
-            (widget.color == null
-                ? buttonTheme.colorScheme.onSurface.withOpacity(disableOpacity)
-                : widget.color.withOpacity(disableOpacity));
-    Color _textColor = widget.enabled
+    effect = enabled ? widget.effect : false;
+    Color backgroundColor = widget.color;
+    if (!enabled) {
+      backgroundColor = widget.disabledColor ??
+          (widget.color ?? Colors.grey[400]).withOpacity(disableOpacity);
+    }
+    Color _textColor = enabled
         ? widget.textColor ?? textStyle.color
         : widget.disabledTextColor ??
             (widget.textColor == null
@@ -358,7 +393,7 @@ class _FButton extends State<FButton> {
                 : widget.textColor.withOpacity(disableOpacity));
 
     /// Handle the relationship between [image] and [loading]
-    widget.image = widget.loading
+    Widget image = loading
         ? SizedBox(
             width: widget.loadingSize,
             height: widget.loadingSize,
@@ -373,9 +408,9 @@ class _FButton extends State<FButton> {
         : widget.image;
 
     /// Handle the relationship between loading text and regular text
-    widget.loadingText =
+    String loadingText =
         widget.loadingText == null ? widget.text : widget.loadingText;
-    widget.text = widget.loading ? widget.loadingText : widget.text;
+    String text = loading ? loadingText : widget.text;
 
     /// Corner
     BorderRadius borderRadius = widget.corner == null
@@ -388,19 +423,19 @@ class _FButton extends State<FButton> {
           );
 
     /// side
-    var sideColor = widget.enabled
-        ? widget.strokeColor ?? Colors.transparent
-        : widget.strokeColor == null
+    Color sideColor = enabled
+        ? shadowColor ?? Colors.transparent
+        : shadowColor == null
             ? Colors.transparent
             : widget.strokeColor.withOpacity(disableOpacity);
-    var borderSide = BorderSide(
+    BorderSide borderSide = BorderSide(
       width: widget.strokeWidth ?? 0,
       color: sideColor,
       style: BorderStyle.solid,
     );
 
     /// shape
-    var shape = widget.cornerStyle == FButtonCornerStyle.round
+    ShapeBorder shape = widget.cornerStyle == FButtonCornerStyle.round
         ? RoundedRectangleBorder(
             borderRadius: borderRadius,
             side: borderSide,
@@ -409,35 +444,35 @@ class _FButton extends State<FButton> {
             borderRadius: borderRadius,
             side: borderSide,
           );
-    var decoration = ShapeDecoration(
-        color: gradientEnable ? null : widget.color,
+    Decoration decoration = ShapeDecoration(
+        color: gradientEnable ? null : backgroundColor,
         gradient: gradientEnable ? widget.gradient : null,
-        shadows: widget.shadowColor != null && widget.shadowBlur != 0
+        shadows: shadowColor != null && shadowBlur != 0
             ? [
                 BoxShadow(
-                  color: widget.shadowColor ??
-                      widget.color.withOpacity(disableOpacity),
+                  color:
+                      shadowColor ?? widget.color.withOpacity(disableOpacity),
                   offset: widget.shadowOffset ?? Offset(0, 0),
                   blurRadius: widget.shadowBlur,
                 )
               ]
             : null,
         shape: shape);
-    var layerText = _buildTextLayer(_textColor, themeData);
-    var layerRow = _buildRowLayer(layerText);
-    var layerContainer = _buildContainerLayer(layerRow);
-    var layerEffect =
+    Widget layerText = _buildTextLayer(text, _textColor, themeData);
+    Widget layerRow = _buildRowLayer(text, image, layerText);
+    Widget layerContainer = _buildContainerLayer(layerRow);
+    Widget layerEffect =
         _buildEffectLayer(_textColor, shape, borderRadius, layerContainer);
-    var layerDecoratedBox = _buildDecoratedBoxLayer(decoration, layerEffect);
-    var layerGestureDetector = widget.effect
+    Widget layerDecoratedBox = _buildDecoratedBoxLayer(decoration, layerEffect);
+    Widget layerGestureDetector = effect
         ? layerDecoratedBox
         : GestureDetector(
-            onTap: widget.enabled ? _onPressed : null,
+            onTap: enabled ? _onPressed : null,
             child: layerDecoratedBox,
           );
-    var result = Semantics(
+    Widget result = Semantics(
       button: true,
-      enabled: widget.enabled,
+      enabled: enabled,
       child: layerGestureDetector,
     );
     return result;
@@ -461,31 +496,34 @@ class _FButton extends State<FButton> {
     return layerDecoratedBox;
   }
 
-  Text _buildTextLayer(Color _textColor, ThemeData themeData) {
+  Text _buildTextLayer(String text, Color _textColor, ThemeData themeData) {
     return Text(
       widget.text ?? "",
-      style: TextStyle(
-          color: _textColor,
-          fontWeight: widget.textWeight,
-          height: widget.fontHeight,
-          fontSize: widget.fontSize ?? themeData.textTheme.button.fontSize,
-          fontStyle: widget.fontStyle ?? themeData.textTheme.button.fontStyle),
+      style: widget.style ??
+          TextStyle(
+              color: _textColor,
+              fontWeight: widget.fontWeight,
+              height: widget.fontHeight,
+              fontSize: widget.fontSize ?? themeData.textTheme.button.fontSize,
+              fontStyle:
+                  widget.fontStyle ?? themeData.textTheme.button.fontStyle),
     );
   }
 
-  Widget _buildRowLayer(Widget layerText) {
-    if (widget.image == null) return layerText;
-    var showLoading = (widget.loading && widget.hideTextOnLoading);
-    if (showLoading || widget.text == null || widget.text == "") {
-      return widget.image;
+  Widget _buildRowLayer(String text, Widget image, Widget layerText) {
+    if (image == null) return layerText;
+    var showLoading = (loading && widget.hideTextOnLoading);
+    if (showLoading || text == null || text == "") {
+      return image;
     } else {
       switch (widget.imageAlignment) {
         case ImageAlignment.left:
           return Row(
             crossAxisAlignment: CrossAxisAlignment.center,
             mainAxisAlignment: MainAxisAlignment.center,
+            mainAxisSize: MainAxisSize.min,
             children: <Widget>[
-              widget.image,
+              image,
               SizedBox(width: widget.imageMargin),
               layerText,
             ],
@@ -494,8 +532,9 @@ class _FButton extends State<FButton> {
           return Column(
             crossAxisAlignment: CrossAxisAlignment.center,
             mainAxisAlignment: MainAxisAlignment.center,
+            mainAxisSize: MainAxisSize.min,
             children: <Widget>[
-              widget.image,
+              image,
               SizedBox(height: widget.imageMargin),
               layerText,
             ],
@@ -504,20 +543,22 @@ class _FButton extends State<FButton> {
           return Row(
             crossAxisAlignment: CrossAxisAlignment.center,
             mainAxisAlignment: MainAxisAlignment.center,
+            mainAxisSize: MainAxisSize.min,
             children: <Widget>[
               layerText,
               SizedBox(width: widget.imageMargin),
-              widget.image,
+              image,
             ],
           );
         case ImageAlignment.bottom:
           return Column(
             crossAxisAlignment: CrossAxisAlignment.center,
             mainAxisAlignment: MainAxisAlignment.center,
+            mainAxisSize: MainAxisSize.min,
             children: <Widget>[
               layerText,
               SizedBox(height: widget.imageMargin),
-              widget.image,
+              image,
             ],
           );
       }
@@ -525,25 +566,24 @@ class _FButton extends State<FButton> {
     return layerText;
   }
 
-  Container _buildContainerLayer(Widget layerRow) {
+  Widget _buildContainerLayer(Widget layerRow) {
     return Container(
       width: widget.width,
       height: widget.height,
-      alignment: Alignment.center,
-      padding: widget.padding ??
-          EdgeInsets.symmetric(horizontal: 16.0, vertical: 16),
+      padding: widget.padding,
+      alignment: widget.alignment,
       child: layerRow,
     );
   }
 
   Widget _buildEffectLayer(Color _textColor, ShapeBorder shapeBorder,
-      BorderRadius borderRadius, Container layerContainer) {
+      BorderRadius borderRadius, Widget layerContainer) {
     return Material(
       clipBehavior: Clip.none,
       type: MaterialType.transparency,
-      child: widget.effect
+      child: effect
           ? InkWell(
-              onTap: widget.enabled ? _onPressed : null,
+              onTap: enabled ? _onPressed : null,
               onHighlightChanged: _handleHighlightChanged,
               onHover: _handleHoveredChanged,
               focusNode: widget.focusNode,
@@ -562,15 +602,15 @@ class _FButton extends State<FButton> {
 
   void _onAnimationEnd() {
     setState(() {
-      widget.shadowBlur = widget.cacheShadowBlur;
-      widget.shadowColor = widget.cacheShadowColor;
+      shadowBlur = cacheShadowBlur;
+      shadowColor = cacheShadowColor;
     });
   }
 
   _onPressed() {
     widget.onPressed();
-    if (widget.clickLoading && !widget.loading) {
-      widget.loading = true;
+    if (widget.clickLoading && !loading) {
+      loading = true;
       setState(() {});
     } else if (widget.clickEffect) {
       _pressOutEffect();
@@ -579,8 +619,8 @@ class _FButton extends State<FButton> {
   }
 
   _pressOutEffect() {
-    widget.shadowBlur = (widget.shadowBlur ?? 0) + 6;
-    widget.shadowColor = widget.shadowColor ?? widget.color.withOpacity(0.68);
+    shadowBlur = (shadowBlur ?? 0) + 6;
+    shadowColor = shadowColor ?? widget.color.withOpacity(0.68);
   }
 }