+++ CHANGELOG.md
@@ -10,3 +10,21 @@
 - add @immutable to fbutton.dart
 
 - minSDK 2.2
+
+## 1.1.0
+
+- Add  **style**  parameter to configure text style. The following parameters will be removed in subsequent versions:
+    - `textColor`
+    - `fontSize`
+    - `fontStyle`
+    - `fontHeight`
+    - `textWeight`
+
+- Add the  **alignment**  parameter to cancel the default centering
+
+- Remove default Padding
+
+- Remove the default background color
+
+- `onPress' is no longer a required parameter
+